refactor: split off mock tests from server collation tests
